fix: conflict 이후 빌드실패 해결
